refactor(transcription): reorganize services by deployment type
- Move transcription services to local/, cloud/, self-hosted/ folders
- Consolidate model constants directly into service files
- Remove constants/transcription folder entirely
- Update all imports to use direct paths instead of barrel exports
- Move transcription-services.ts to services/transcription/registry.ts
- Create shared LocalModelConfig type in types/models.ts
