+++ apps/whispering/src/lib/components/settings/selectors/TranscriptionSelector.svelte
@@ -8,7 +8,7 @@
 	import {
 		TRANSCRIPTION_SERVICES,
 		type TranscriptionService,
-	} from '$lib/constants/transcription';
+	} from '$lib/services/transcription/registry';
 	import {
 		getSelectedTranscriptionService,
 		isTranscriptionServiceConfigured,
