--- apps/whispering/src/lib/constants/transcription/index.ts
@@ -1,23 +1,0 @@
-export {
-	OPENAI_TRANSCRIPTION_MODELS,
-	type OpenAIModel,
-} from '$lib/services/transcription/openai';
-
-export { GROQ_MODELS, type GroqModel } from '$lib/services/transcription/groq';
-
-export {
-	ELEVENLABS_TRANSCRIPTION_MODELS,
-	type ElevenLabsModel,
-} from '$lib/services/transcription/elevenlabs';
-
-export {
-	DEEPGRAM_TRANSCRIPTION_MODELS,
-	type DeepgramModel,
-} from '$lib/services/transcription/deepgram';
-
-export {
-	TRANSCRIPTION_SERVICE_IDS,
-	TRANSCRIPTION_SERVICES,
-	TRANSCRIPTION_SERVICE_OPTIONS,
-	type TranscriptionService,
-} from './transcription-services';
