+++ apps/whispering/src/lib/services/transcription/cloud/groq.ts
