+++ apps/whispering/src/lib/services/transcription/cloud/openai.ts
