+++ apps/whispering/src/lib/services/transcription/index.ts
@@ -1,25 +1,24 @@
-import { DeepgramTranscriptionServiceLive } from './deepgram';
-import { ElevenlabsTranscriptionServiceLive } from './elevenlabs';
-import { GroqTranscriptionServiceLive } from './groq';
-import { OpenaiTranscriptionServiceLive } from './openai';
-import { SpeachesTranscriptionServiceLive } from './speaches';
-import { WhisperCppTranscriptionServiceLive } from './whispercpp';
-import { ParakeetTranscriptionServiceLive } from './parakeet';
+// Direct imports and re-exports from organized services
+
+// Local transcription services
+import { WhisperCppTranscriptionServiceLive } from './local/whispercpp';
+import { ParakeetTranscriptionServiceLive } from './local/parakeet';
+
+// Cloud transcription services
+import { DeepgramTranscriptionServiceLive } from './cloud/deepgram';
+import { ElevenlabsTranscriptionServiceLive } from './cloud/elevenlabs';
+import { GroqTranscriptionServiceLive } from './cloud/groq';
+import { OpenaiTranscriptionServiceLive } from './cloud/openai';
+
+// Self-hosted transcription services
+import { SpeachesTranscriptionServiceLive } from './self-hosted/speaches';
 
 export {
+	WhisperCppTranscriptionServiceLive as whispercpp,
+	ParakeetTranscriptionServiceLive as parakeet,
+	DeepgramTranscriptionServiceLive as deepgram,
 	ElevenlabsTranscriptionServiceLive as elevenlabs,
 	GroqTranscriptionServiceLive as groq,
 	OpenaiTranscriptionServiceLive as openai,
 	SpeachesTranscriptionServiceLive as speaches,
-	DeepgramTranscriptionServiceLive as deepgram,
-	WhisperCppTranscriptionServiceLive as whispercpp,
-	ParakeetTranscriptionServiceLive as parakeet,
 };
-
-export type { ElevenLabsTranscriptionService } from './elevenlabs';
-export type { GroqTranscriptionService } from './groq';
-export type { OpenaiTranscriptionService } from './openai';
-export type { SpeachesTranscriptionService } from './speaches';
-export type { DeepgramTranscriptionService } from './deepgram';
-export type { WhisperCppTranscriptionService } from './whispercpp';
-export type { ParakeetTranscriptionService } from './parakeet';
