+++ apps/whispering/src/lib/services/transcription/self-hosted/speaches.ts
